+++ executor_nocoverage.go
@@ -1,3 +1,0 @@
-package dogma
-
-func (IdempotencyKeyOption) isExecuteCommandOption() {}